Make sure arrays/lists passed to logtrapz are floats

+++ lintegrate/lintegrate.pyx
@@ -140,17 +140,17 @@ def logtrapz(f, x, disable_checks=False, args=()):
 
             if not disable_checks:
                 # make sure x values are in ascending order (keeping f values associated to their x evaluation points)
-                zp = np.array(sorted(zip(x, f)))
+                zp = np.array(sorted(zip(x, f)), dtype=np.float64)
 
                 # perform trapezium rule (internal logtrapzC function is faster than using scipy logsumexp)
                 return logtrapzC(zp[:,1], zp[:,0])
             else:
-                return logtrapzC(np.array(f), np.array(x))
+                return logtrapzC(np.array(f, dtype=np.float64), np.array(x, dtype=np.float64))
         elif isinstance(x, float):
             assert x > 0., "Evaluation spacings must be positive"
 
             # perform trapezium rule
-            return logtrapzC(f, np.array([x]))
+            return logtrapzC(f.astype(np.float64), np.array([x]))
         else:
             raise TypeError('Error... value of "x" must be a numpy array or a float')
     elif callable(f): # f is a function
@@ -161,18 +161,18 @@ def logtrapz(f, x, disable_checks=False, args=()):
             try:
                 if not isinstance(args, tuple):
                     args = (args,)
-                vs = f(np.array(x), args) # make sure x is an array when passed to function
+                vs = f(np.array(x, dtype=np.float64), args) # make sure x is an array when passed to function
             except Exception as e:
                 raise RuntimeError('Error... could not evaluate function "f": {}'.format(e))
 
             if not disable_checks:
                 # make sure x values are in ascending order (keeping f values associated to their x evaluation points)
-                zp = np.array(sorted(zip(x, vs)))
+                zp = np.array(sorted(zip(x, vs)), dtype=np.float64)
 
                 # perform trapezium rule (internal logtrapzC function is faster than using scipy logsumexp)
                 return logtrapzC(zp[:,1], zp[:,0])
             else:
-                return logtrapzC(vs, np.array(x))
+                return logtrapzC(vs, np.array(x, dtype=np.float64))
         else:
             raise TypeError('Error... "x" must be a numpy array or list')
     else:
